Rename state::impls::nodes to state::impls::local_node
